+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 7.11
  - Added hccapx support for hashcat 3.40 or higher
  - Decrypted hashcat trophy files beautified
+ - Added confirm questions before attacks if no target selected
 
 ### 7.1
  - Added default choice for some yes/no question. Regexp improved